refactor(core): convert the ExecutionRoot.vue properly
